Upgrade to fuse3

+++ abstract_storage.h
@@ -67,7 +67,7 @@ public:
      * @brief size - return size of data currently in storage
      * @return size of underlying data in bytes
      */
-    virtual size_t size() = 0;
+    virtual off_t size() = 0;
 
     /**
      * @brief read - read from storage to buffer @ref buf
